Melhora na documentação da page Flashcard e adição de quantidade de cards estudados
